+++ packages/auto_updater/lib/src/updater_listener.dart
@@ -1,4 +1,5 @@
-import 'package:auto_updater/auto_updater.dart';
+import 'package:auto_updater/src/appcast.dart';
+import 'package:auto_updater/src/updater_error.dart';
 
 abstract mixin class UpdaterListener {
   void onUpdaterError(UpdaterError? error);
@@ -7,4 +8,13 @@ abstract mixin class UpdaterListener {
   void onUpdaterUpdateNotAvailable(UpdaterError? error);
   void onUpdaterUpdateDownloaded(AppcastItem? appcastItem);
   void onUpdaterBeforeQuitForUpdate(AppcastItem? appcastItem);
+  
+  /// Called when a mandatory update is available
+  void onUpdaterMandatoryUpdateAvailable(AppcastItem? appcastItem) {}
+  
+  /// Called when update installation begins (for mandatory updates)
+  void onUpdaterUpdateInstallationStarted(AppcastItem? appcastItem) {}
+  
+  /// Called when update installation is completed
+  void onUpdaterUpdateInstallationCompleted(AppcastItem? appcastItem) {}
 }